requirements.txt
PyYAML
pyautogui
clipboard
Jinja2
PyPDF2
Pillow
